Split fleet observation match tests into a dedicated module.
Keeps ingest and arbitration suites under the 1k-line maintainability
threshold without changing production behavior.
